Add format flag to get_port_conf
